% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retel-package.R
\docType{package}
\name{retel-package}
\alias{retel-package}
\title{retel: Regularized Exponentially Tilted Empirical Likelihood}
\description{
Implements the regularized exponentially tilted empirical likelihood method. Details of the method are given in Kim, MacEachern, and Peruggia (2023) \doi{10.48550/arXiv.2312.17015}. This work was supported by the U.S. National Science Foundation under Grants No. SES-1921523 and DMS-2015552.
}
\references{
Kim E, MacEachern SN, Peruggia M (2023).
"Regularized Exponentially Tilted Empirical Likelihood for Bayesian
Inference."
\doi{10.48550/arXiv.2312.17015}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markean/retel}
  \item Report bugs at \url{https://github.com/markean/retel/issues}
}

}
\author{
\strong{Maintainer}: Eunseop Kim \email{markean@pm.me} (\href{https://orcid.org/0009-0000-2138-788X}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Steven MacEachern [contributor, thesis advisor]
  \item Mario Peruggia [contributor, thesis advisor]
}

}
\keyword{internal}
