% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retel.R
\name{retel}
\alias{retel}
\title{Regularized exponentially tilted empirical likelihood}
\usage{
retel(fn, x, par, mu, Sigma, tau, opts)
}
\arguments{
\item{fn}{An estimating function that takes the data \code{x} and parameter value
\code{par} as its arguments, returning a numeric matrix. Each row is the return
value from the corresponding row in \code{x}.}

\item{x}{A numeric matrix, or an object that can be coerced to a numeric
matrix. Each row corresponds to an observation. The number of rows must be
greater than the number of columns.}

\item{par}{A numeric vector of parameter values to be tested. The length of
the vector must be the same as the number of columns in \code{x}.}

\item{mu}{A numeric matrix.}

\item{Sigma}{A numeric matrix.}

\item{tau}{A single numeric.}

\item{opts}{A list with optimization options for \code{\link[=nloptr]{nloptr()}}.}
}
\value{
A single numeric of the log-likelihood ratio.
}
\description{
Computes regularized exponentially tilted empirical likelihood.
}
\examples{
set.seed(63456)
f <- function(x, par) {
  x - par
}
x <- rnorm(100)
par <- 0
mu <- 0
Sigma <- matrix(rnorm(1), nrow = 1)
tau <- 1
retel(f, x, par, mu, Sigma, tau)
}
\references{
Kim E, MacEachern SN, Peruggia M (2023).
"Regularized Exponentially Tilted Empirical Likelihood for Bayesian
Inference."
\doi{10.48550/arXiv.2312.17015}.
}
