% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etel.R
\name{etel}
\alias{etel}
\title{Exponentially tilted empirical likelihood}
\usage{
etel(g)
}
\arguments{
\item{g}{A numeric matrix, or an object that can be coerced to a numeric
matrix. Each row corresponds to an observation of an estimating function.
The number of rows must be greater than the number of columns.}
}
\value{
An object of class?.
}
\description{
Computes exponentially tilted empirical likelihood.
}
\references{
Schennach, SM (2005).
"Bayesian exponentially tilted empirical likelihood"
\emph{Biometrika}, 92, 31--46.
}
